separated database connection and express app instance so testing will work
